+++ packages/bluesky/lib/bluesky.dart
@@ -138,6 +138,7 @@ export 'package:atproto_core/atproto_core.dart'
         HttpMethod,
         HttpStatus,
         Serializable,
+        Platform,
         AtUri,
         NSID,
         CID,